fix noredmine

+++ spring-data-jdbc/src/main/java/org/springframework/data/jdbc/core/mapping/BasicJdbcPersistentProperty.java
@@ -60,6 +60,6 @@ public class BasicJdbcPersistentProperty extends BasicRelationalPersistentProper
 
 	@Override
 	public int hashCode() {
-		return (int) getProperty().hashCode() * getOwner().hashCode();
+		return getProperty().hashCode() * getOwner().hashCode();
 	}
 }
